requirements.txt
vapi_server_sdk>=0.8.1
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.3
pydantic-settings>=2.0.0
typing-extensions>=4.10.0
starlette>=0.36.3
loguru>=0.7.2
pyngrok>=7.1.5
httpx>=0.21.2
pyaudio>=0.2.13
requests>=2.31.0
vapi-python==0.1.6
dotenv>=1.0.1